+++ client/src/common/fileManager/FileManagerTable.js
@@ -16,6 +16,7 @@
 import React, {useContext, useState} from 'react';
 import {
   Download,
+  Edit,
   File,
   Folder,
   FolderPlus,
@@ -29,6 +30,8 @@ import {Dropdown} from "react-bootstrap";
 import PropTypes from "prop-types";
 import NotyfContext from "../../context/NotyfContext";
 import FileManagerAddToStudyModal from "./FileManagerAddToStudyModal";
+import FileManagerRenameFolderModal from "./FileManagerRenameFolderModal";
+import axios from "axios";
 
 const FileManagerTable = ({
   folder,
@@ -41,6 +44,7 @@ const FileManagerTable = ({
   const [addToAssayModalIsOpen, setAddToAssayModalIsOpen] = useState(false);
   const [addToProgramModalIsOpen, setAddToProgramModalIsOpen] = useState(false);
   const [selectedItem, setSelectedItem] = useState(null);
+  const [renameFolderModalIsOpen, setRenameFolderModalIsOpen] = useState(false);
 
   const handleItemClick = (item, index) => {
     setSelectedItem(item);
@@ -58,6 +62,28 @@ const FileManagerTable = ({
   const handleCopyUrl = (d) => {
     navigator.clipboard.writeText(d.url);
     notyf.open({message: "Copied URL clipboard", type: "success"});
+  }
+
+  const handleRenameFolder = (values, { setSubmitting, resetForm }) => {
+    setSubmitting(true);
+    const data = new FormData();
+    data.append("folderId", rootFolder.id);
+    data.append("path", values.path);
+    data.append("name", values.name);
+    axios.post("/api/internal/data-files/rename-folder", data)
+    .then(() => {
+      notyf.open({message: "Folder renamed successfully.", type: "success"});
+    })
+    .catch(e => {
+      console.error(e);
+      console.error("Failed to rename folder");
+      notyf.open({message: "Failed to rename folder. Please try again.", type: "error"});
+    })
+    .finally(() => {
+      setSubmitting(false);
+      resetForm();
+      setRenameFolderModalIsOpen(false);
+    });
   }
 
   const columns = [
@@ -175,6 +201,19 @@ const FileManagerTable = ({
                   d.type === "folder" && (
                     <>
 
+                      {
+                        rootFolder.storageDrive && rootFolder.storageDrive.driveType
+                        && ["ONEDRIVE", "LOCAL"].includes(rootFolder.storageDrive.driveType) ? (
+                            <Dropdown.Item onClick={() => {
+                                setSelectedItem(d);
+                                setRenameFolderModalIsOpen(true);
+                            }}>
+                              <Edit className="align-middle me-2" size={18} />
+                              Rename Folder
+                            </Dropdown.Item>
+                        ) : ""
+                      }
+
                       <Dropdown.Item onClick={() => {
                         setSelectedItem(d);
                         setAddToProgramModalIsOpen(true);
@@ -290,6 +329,13 @@ const FileManagerTable = ({
         folder={selectedItem}
         rootFolder={rootFolder}
         type={"assay"}
+      />
+
+      <FileManagerRenameFolderModal
+          folder={selectedItem}
+          setModalIsOpen={setRenameFolderModalIsOpen}
+          isOpen={renameFolderModalIsOpen}
+          handleFormSubmit={handleRenameFolder}
       />
 
     </>